🧹 cleanup of game files and deleted route example
🗑️ deleted the route example for leaderboard
🧹 put the game files in the route folder as it is a route
