--- front/src/routes/GamePage.tsx
@@ -1,11 +1,0 @@
-import Game from "../Game";
-
-export default function GamePage() {
-    return (
-      <div>
-        {/* <main style={{ padding: "1rem 0" }}> */}
-        <Game></Game>
-      {/* </main> */}
-      </div>
-    );
-  }
